day5/requirements.txt
torch
transformers
sentence-transformers
mlflow
openai
requests
pyyaml
